+++ aps/wavepy2/tools/imaging/single_grating/widgets/curv_from_height_widget.py
@@ -96,5 +96,5 @@ class CurvFromHeight(WavePyWidget):
 
         layout.addWidget(FigureCanvas(figure1))
 
-        self.setFixedWidth(plot_tools.WIDGET_FIXED_WIDTH * 1.4)
+        self.setFixedWidth(int(plot_tools.WIDGET_FIXED_WIDTH * 1.4))
         self.setFixedHeight(700)
